+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [x.x.x] - Unreleased
 
+### Added
+- `Proof` model and `ProofType` enum
+- `proof` property on rows returned by `getSheet` and `getReport`
+- `PROOFS` inclusion value for `SheetInclusion`, `ReportInclusion`, and `RowInclusion`
+
 ## [4.3.0] - 2026-07-20
 
 ### Added