+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [x.x.x] - Unreleased
 
+## [4.4.0] - 2026-08-12
+
 ### Added
 - `Proof` model and `ProofType` enum
 - `proof` property on rows returned by `getSheet` and `getReport`